fix/탐색페이지 코드 파일분리,탐색페이지 카드크기 조정 및 dvh->svh로 수정
