+++ src/features/explore-cards/index.ts
@@ -1,0 +1,2 @@
+export { useExploreCards } from './model/useExploreCards';
+export { ExploreCardArea } from './ui/ExploreCardArea';
